Changes for RTLAZY
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@886 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/CRUN/CheckReferenceAssignment.c
@@ -27,7 +27,7 @@ void ChkRA()
 #ifdef RTLAZY
     /* It may be a dangling (negative) reference */
     if (isLazyRef(*theObjHandle))
-      negAOArefsINSERT(theObjHandle);
+      negAOArefsINSERT((long) theObjHandle);
 #endif
     return; 
   }
